requirements.txt: install 'shapely' from source
As mentioned in Cartopy's installation guide [1], Shapely must be
installed from source in order to link to the same GEOS library as
Cartopy, since the Shapely wheel includes its own version of GEOS.

Using the wheel version results in a runtime failure (probably due to
incompatible and different versions of GEOS being used by Cartopy and
Shapely) and this error message (and Python segmentation fault!):

    Geometry must be a Point or LineString
    python: geos_ts_c.cpp:4179: int GEOSCoordSeq_getSize_r(GEOSContextHandle_t, const geos::geom::CoordinateSequence*, unsigned int*): Assertion0 != cs' failed`.
    Aborted (core dumped)

Add the '--no-binary' flag to the requirements.txt for Shapely, in
accordance to the Requirements File Format [2], to instruct Pip to
always install Sahpely from source.

Note that this is not an optimal solution as the next 'pip freeze' will
remove that flag.

[1] https://scitools.org.uk/cartopy/docs/latest/installing.html#required-dependencies
[2] https://pip.pypa.io/en/stable/cli/pip_install/#requirements-file-format