README: manually download the Cartopy shapefiles
The URL from where the Cartopy shapefiles are downloaded is out of
service [1], and the new Cartopy vesion that corrects the URL (0.20)
requires Python 3.7.

To keep environment changes to a minimum, add instructions to download a
script from the Cartopy repository and run that script to manually
download the shapefiles from the new URL.

The alternative would be to ditch the system Python and use conda to
isntall Cartopy 0.20 (and a more recent Python), but this change is more
minimalist.

[1] https://github.com/SciTools/cartopy/pull/1833